Preserve the name of a staticmethod if the decorator doesn't change its signature.
Summary:
Problem: When applying a decorator on a function, we preserve the function's name if the type signature is unchanged. This leads to better error messages (i.e., "In call foo" instead of "In anonymous call").

However, when we have a staticmethod that is decorated, we always lose the function name.

Reasoning for why we preserve the name only when the type signature is preserved: D21487479.
It basically said that we want to preserve the same type for a given global name.

Reviewed By: grievejia

Differential Revision: D46942109

fbshipit-source-id: 2e4eade249ff889f77dfdf941569345aa32d1f4a