Add functionality for closure reads in inner functions
Summary:
The main changes are in forwardAnalysis and backwardAnalysis.

Now in backward analysis, both captures (variables which are accessed in the function, but not passed in, also known as closures), and positional parameters are treated as formal parameters in the resulting sink model.

In forward analysis, the closures from the sink model are joined with the current state at the evaluation of a function call. Specifically the joining matches based on the fully qualified name, which has the closure information.

The result is now taint models contain nonlocals.

Reviewed By: arthaud

Differential Revision: D46856471

fbshipit-source-id: 60815982b1d47ff9d5232f8740833b0429c702fc