Split daemon querier and remote index call hierarchy handlers
Summary:
This diff splits the call hierarchy logic into `prepare_call_hierarchy` (initial call hierarchy request) and `get_call_hierarchy_from_item` (drilling down/up the hierarchy).

This enables us to lazily compute the fan out for each symbol as necessary.

This diff focuses primarily on the refactoring and splitting of existing logic. New logic will be implemented in the next diff

The Implementation for following functions are to be updated in the next diff
- `glass_prepare_call_hierarchy`
- `glass_get_call_hierarchy_from_item`
- `call_hierarchy_from_item`

Reviewed By: stroxler

Differential Revision: D47035536

fbshipit-source-id: fee0781a1bf4a0d3aa4c40f679dd5f19e2399fe3