Implement incoming/outgoing call hierarchy request and response structures in protocol
Summary:
## Overview
The call hierarchy request is sent from the client to the server to return a call hierarchy for the language element of given text document positions. The call hierarchy requests are executed in two steps:

1. A call hierarchy item is resolved for the given text document position (this was previous implemented)
2. For a call hierarchy item the incoming or outgoing call hierarchy items are resolved. This event is triggered once following the initial prepareHierarchy request and once each time a user interacts and selects the dropdown on the call hierarchy panel.

To support this, the following diffs (10 - 14) will first add support for the  [IncomingCall](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.18/specification/#callHierarchy_incomingCalls) (caller hierarchy) and [OutgoingCall](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.18/specification/#callHierarchy_outgoingCalls) (callee hierarchy) requests, and then split the existing logic for slight performance gains.

This diff also introduces `PyreCallHierarchyRelationDirection` which is used in the next diff to define a general function that supports both caller and callee hierarchy (without coupling the LSP to the Glass service types)

## Example:
Consider invoking "Show Call Hierarchy" on `a()` which is called by `b()` which is in turn called by `c()` and `d()`. In this example we only consider following the caller hierarchy.

 {F1037750701}

This will trigger the following series of request/response exchange
- Clicking "Show Call Hierarchy" sends a `textDocument/prepareCallHierarchy`
    - language server responds with a CallHierarchyItem that represents `a()`. (More specifically, we return an array of CallHierarchyItems but the LSP client only expects one item in the array.)
```
[
   CallHierarchyItem(
      name='a',
      kind=FUNCTION
   )
]
```
- Toggling the dropdown on `a()` sends `callHierarchy/incomingCalls`
    - language server responds with a `CallHierarchyIncomingCall` that contains a CallHierarchyItem that represents `b()`
```
CallHierarchyIncomingCall(
   items=[
       CallHierarchyItem(
           name='b',
           kind=FUNCTION
       )
   ]
)
```

- Toggling the dropdown on `b()` sends `callHierarchy/incomingCalls`
    - language server responds with a `CallHierarchyIncomingCall` that contains a CallHierarchyItem that represents `c()` and `d()`
```
CallHierarchyIncomingCall(
   items=[
       CallHierarchyItem(
           name='c',
           kind=FUNCTION
       ),
       CallHierarchyItem(
           name='d',
           kind=FUNCTION
       )
   ]
)
```

## Implemented Demo
 https://pxl.cl/2RvTh

Reviewed By: stroxler

Differential Revision: D46896877

fbshipit-source-id: 7d3c1ec6d92d51eef943679d1292201484edc5d4