Add telemetry logging for call hierarchy related events
Summary:
We add telemetry logging to gather usage statistics for this new feature. We log standard information about the input arguments and responses. In particular we want to know how often our responses are non-empty (`non_empty`), and how long each the responses take (`duration_ms`).

Following this update, we intended to add a dashboard
- Was this call hierarchy response empty?
- How long did it take? p50, p90, p99

Reviewed By: kinto0

Differential Revision: D47069174

fbshipit-source-id: 9f585968291458afd1fcb61250289e3312426cbb