Back prop for turning captures to formals
Summary:
This work allows for `captured_variables` to change to `formals` in the back propagation for the backward analysis to match the forward analysis that extracts captures from taint models.

I need some help with the `hof` portion, however. I tried to reason through what is simulated for the `hof` part, but am unsure what to do with that simulation and the captures to find the correct flow.

Reviewed By: arthaud

Differential Revision: D47124368

fbshipit-source-id: eea25e168d23c617d20691bccd6bb5c15189036e