Refactor the function that extracts type annotations for a modeled parameter
Summary:
See D47097963 for context.

In the following diffs, we will need to look into the type annotation of a parameter modeled in a `.pysa` file in order to implement `.all_static_fields()`. Luckily, we were already fetching the annotation of a modeled parameter to compute the type breadcrumbs.
This diff extracts that functionatily into a separate function so it can be used in the next diff.

Reviewed By: tianhan0

Differential Revision: D47131763

fbshipit-source-id: f2b30588b4c71f9002044011468660ed3ba65c30