Improve accuracy with refined go-to-def criteria
Summary: Following analysis and discussion of the performance (notes [found here](https://fb.workplace.com/notes/267666129282519)) we noticed certain inconsistent states arising. To gain some preliminary understanding of the issue, we learn towards accuracy first only using the indexed result when we are certain that the Pyre client is unable to respond i.e. is disconnected

Reviewed By: kinto0, pradeep90

Differential Revision: D47136971

fbshipit-source-id: f880090b421e1f872f266e9129cdd60e6d088087