Fix error when printing differences on test errors
Summary:
When introducing an error in our integration tests and running our tests with `make test`, we now see the following error:
```
/bin/sh: -c: line 0: syntax error near unexpected token `('
/bin/sh: -c: line 0: `diff -u <(printf '[REDACTED]') <(printf '[REDACTED]')
```
It looks like `<(cmd)` does not work with `/bin/sh`, but it does with `/bin/bash`.

This diff fixes the test library to use `/bin/bash` instead.
The only downside is that it probably doesn't work on Windows - although I haven't tested and I think we use a lot of other Unix* stuff that doesn't work.

Reviewed By: grievejia

Differential Revision: D47155684

fbshipit-source-id: 751ef3970a388b1f14ef4d04094f46c440385278