add client-version to pyre cli
Summary:
the pyre client has a `--version` flag which downloads the binary then displays the binary version + client version. when the binary version is unnecessary, this downloads the binary which may take ~10 seconds.

we would like a hook to see the client version in without downloading the binary in the vscode extension to add to logs. this diff adds a --version argument to do that while keeping the behavior backwards compatible.

Reviewed By: grievejia

Differential Revision: D47378711

fbshipit-source-id: 1374a0137a3641caf3b5f8aa00a8bd464fbf217e