Immediate flush of ERRPY telemetry
Summary:
It looks like the current mechanism by which we flush logs to scuba is a bit too conservative for ERRPY - 500 lines or until 6 hours has passed - so lets just flush everything for ERRPY.

Without flushing, I believe that we can loose data if pyre is dirt shut down (which I think  occurs when an on-demand is released as I got no logs passed through to scuba when I did this). Hence I think that in the future we need to revise the way in which we flush logs, perhaps flushing every 10 minutes is a good compromise (and when we are on OCaml 5 - which has concurrency, we can enforce this on a timer as opposed to on event).

Anyway, this solves the problem for ERRPY for now

Reviewed By: grievejia

Differential Revision: D47446016

fbshipit-source-id: 3a7176e422e6c8ee0a1621ebd330edfbf7705a9d