Fix opam error in the setup script
Summary:
This diff fixes the opam error `opam: --assume-depexts was added in version 2.1 of the opam CLI, but version 2.0 has been requested, which is older.`.
We need to provide `--cli=2.1`, otherwise opam might decide to run with the 2.0 interface which does not have `--assume-depexts`.
This required a small refactoring of the `Setup` class. In order to keep it stateless, I moved the `detect_opam_version` logic outside of it.

Reviewed By: stroxler

Differential Revision: D47516645

fbshipit-source-id: 703c9117b5874bde3a6ebe23704a6c50df8f3c9e