Fix: derived classes of frozen must be frozen tools/pyre/client/commands/pyre_language_server.py
Summary:
Meta is moving to Python 3.10 for a faster, safer, richer Python experience.

In Python 3.10, the following is an error:
```
dataclass(frozen=True)
class Base:
  pass

class Child(Base):
  pass
```
The fix is to make `Child` frozen as well.

We need to fix this anti-pattern everywhere it appears in order to unblock Python 3.10.

This diff makes this fix to 1 files.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: grievejia

Differential Revision: D47702830

fbshipit-source-id: 3c96ee99bb3f939ff232f8a27fc7e807171a8fa6