Flip client to using `xxx` rather than `newxxx`
Summary:
In 2021 we had an old-to-new migration of backend commands, where
we rewrote all of the client-server interactions so that buck
integration could happen server-side.

The v2 backend commands were all named "newxxxx", e.g. "newcheck"
rather than "check", so that the old vs new commands could be
distinguished.

Today, the original commands are gone, and we added entrypoints
so that the old and new names map to the same ocaml entrypoints
over 6 months ago (which means that 100% of binaries running
now support the regular names).

It's time to cut the client over to use the regular names. After
we bump the client in every project we can then get rid of the
`newxxx` names on the backend.

Reviewed By: inseokhwang

Differential Revision: D47726961

fbshipit-source-id: 2026cec129daab911fc34dd5e297607247fd399a