Call smart constructor when instantiating `ReadOnly[T]`.
Summary:
Problem: We use `ReadOnly.create` to create readonly types (and skip allowlisted types). But the `Transformer` module directly creates `ReadOnly Foo`. This leads to problems when instantiating `ReadOnly[T]` with one of the allowlisted types.

Solution: Make it call the constructor.

To break the dependency cycle between the `Transformer` and `ReadOnly` modules, I pulled up `create_readonly` so it could be used by both.

Not sure if there's a way to hide the constructor (`private`) for just one of the variants in an ADT. We'd probably have to mark all the variants as private, which would be a big change.

Reviewed By: stroxler

Differential Revision: D47419117

fbshipit-source-id: ad44898702aacf36b866f02045f4cd9789185476