Add a breadcrumb for model simplification broadening
Summary:
We currently add some breadcrumbs when we broaden (i.e collapse) taint for performance purposes.
One of those breadcrumbs is `widen-broadening`, which is supossedly added when a taint tree is collapsed during widening, which can happen in loops.
The problem is that this breadcrumb is also used when doing model simplification, which is the step that applies our heuristics and limit the size of our models (see https://pyre-check.org/docs/pysa-advanced/#tune-the-taint-tree-width-and-depth).
This diff proposes to introduce a new breadcrumb, `model-broadening` which is added for that special case. The `widen-broadening` will still be added during widening (i.e, loops).

Reviewed By: r0rshark

Differential Revision: D47770341

fbshipit-source-id: bad45c0e16fcf22c65653117ffd9c2341a87ee09