Refactor `List.rev` into a function that takes argument_matches
Summary:
To prevent bugs like the one fixed in the previous diff, let's change the function to be a wrapper over the fold and reverse.

We need a `List.rev` since when folding over and rebuilding the List efficiency with front of list appends, the list is automatically reversed.

Reviewed By: arthaud

Differential Revision:
D47835415

Privacy Context Container: L1152058

fbshipit-source-id: cbeae55fc6cc4afd2c0bfc0de6d99a6c4763d487