Add neccessary reverse for parameters to fix parameter matching bug
Summary:
For the same reason we need to reverse the `argument_taint`, we need to reverse the `capture_taint` so it matches up with captures.

The tests previously only had 1 capture, so this wasn't noticed. Now I even added a 3 parameter test where only 2 are swapped.

Reviewed By: arthaud

Differential Revision:
D47835417

Privacy Context Container: L1152058

fbshipit-source-id: 221a3d49af6705a49771d5eb7a2c87cc5b5067d7