Rename `callable` into `definition`
Summary:
# Problem
We often use the word `callable` in our codebase, to represent the name of a function or method, or in general, anything object that can be called.
However, we often want to deal with specific kind of callables:
* callables with a explicit definition, i.e a `def foo():` statement
* stub callables, defined with the ellipsis `def foo(): ...`
* other callables, i.e function or methods that can be called but don't have a `def foo():` statement. For instance, generated methods such as the `__init__` method for dataclasses (see previous diff).

Hence, using `callable` can be confusing depending on the context.
Most specifically, the `FetchCallables` module - which is responsible for finding all "callables" in the codebase - suffers a lot from this problem. For instance, it has a `get_non_stub_callables` which does not actually return all callables, but only definitions.

# Solution
To avoid all these confusions, let's use the term `definition`, i.e a callable with a define statement (`def foo():`).
This diff renames `callables` into `definitions` in all the places where that makes sense.
Note that pyre itself uses `defines` or `definitions`, rather than `callable`.

Reviewed By: grievejia

Differential Revision: D47874157

fbshipit-source-id: bc514ed812d6a3e57d938a61700937a8ce72accb