Back out "Immediate flush of ERRPY telemetry"
Summary: This change backs out the change to more quickly flush ERRPY telemetry. Note that this does not fix the underlying bug which this change had surfaced, but landing this diff will mean that if we upgrade the pyre server used by codenav (accidentally or deliberately as in D47814753), codenav won't start failing at an unacceptable rate as it was before again.

Reviewed By: grievejia

Differential Revision: D47973365

fbshipit-source-id: 9ffae92bedd5fc0d867d63c46cb5dd459ba7af04