Make --dot-pyre-directory flag available in open source version
Summary:
This diff is in response to a [GitHub issue](https://github.com/facebook/pyre-check/issues/768) that was opened last week.

The user requested a way to move the .pyre directory outside of their project. This diff makes the `--dot-pyre-directory` flag unhidden for non-internal versions of Pyre, so that users can use it in their configurations.

Reviewed By: grievejia

Differential Revision: D47975577

fbshipit-source-id: 4d62c93a348b6fbe8bf6fe7f1ca1b5e6f979912a