Display position in TransformDuplicate errors
Previously when TransformDuplicate errors occured, we just left a note
specifying the duplicate name of the transform. Now we can specify the
locations where the duplicates occur.

Modifies taintTransform to have a new type that holds location
information (for parsed transforms, it is made None). The location
information will only be present for transform node definitions in
taint.config.

Modifies and updates tests to confirm and check the new type of
TransformDuplicate error.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>