Pretty print the via value port in our json output
Summary:
We shouldn't leak the internal representation of ports in our json.
Let's use the same format as callee port for `ViaValueOf` and `ViaTypeOf` parameters.

In the same diff, we explicit the different representation of roots with `pp/show_internal` and `pp/show_external`, similar to our API for targets.

Reviewed By: tianhan0

Differential Revision: D48195342

fbshipit-source-id: 771c15290f426be13ca3fc147f03b667e726add8