Factor out the pattern match on `X[Y]` in models
Summary:
The taint annotation parser pattern matches on the 2 following patterns: `X[Y]` or `X`.
However, that pattern match is duplicated in a lot of places.
This diff refactors the parsing of taint annotations so we pattern match on `X[Y]` and then match on the actual `X` with a nested match.
In my opinion, this makes it a lot easier to read the parsing logic.

Reviewed By: dkgi

Differential Revision: D48190794

fbshipit-source-id: 44026fa2b940e3afc4e8c6714a4215859b9b3196