Use a new module to support storing the override graph in the shared memory
Summary:
Currently, we simply use `SharedMemory.FirstClass` to store the
override graph in the shared memory, which however does not support removing
the override graph from the shared memory -- this is needed when in the next
run, we detect that the old override graph is stale. It needs to be removed, so
that it won't pollute the new override graph.

The removal is not supported because it requries partially invalidating the
shared memory, which requires remembering which keys in the shared memory
(which is a hash table) need to be removed. This functionality will be added in
the next diff in the new module.

Reviewed By: arthaud

Differential Revision: D48078621

fbshipit-source-id: 4d03beb86672c8d56efa1e68508864f236a81a6e