Introduce a new taint annotation `ViaAttributeName` for attribute models
Summary:
# Problem
There are cases where we want to know precisely which attribute of a class was tainted.

For instance:
```
class Person:
  first_name: str
  last_name: str

  def __init__(self, first_name, last_name):
    self.first_name = first_name
    self.last_name = last_name

def to_sink(x):
  sink(Person(first_name="foo", last_name=x))
```
Here we would like to know that `last_name` was tainted, and have that information on the issue.

# Existing solution
While we do have `first-field` and `first-index` breadcrumbs, those do not work here since they are not inferred on taint-in-taint-out for `Person.__init__`. Inferring `first-field` on tito would likely blow up the analysis.

# Solution
We introduce a new taint annotation `ViaAttributeName` for attributes, which simply adds a breadcrumb of the form `via-attribute:<name>` on flows through that attribute. This can then be added on a large set of attributes using a model query. For instance:
```
ModelQuery(
  find="attributes",
  where=[cls.name.equals("Person")],
  model=[AttributeModel(ViaAttributeName)]
)
```

Note this provides very good precision, this can lead to a significant increase in analysis time.

Reviewed By: alexkassil

Differential Revision: D48240710

fbshipit-source-id: 3f80a2bec420615850dce3c4b8022c3ddd800554