Fix `ViaTypeOf` and `ViaAttributeName` on dataclass attributes
Summary:
Since we do not see the implementation of the constructor for dataclasses, we do not properly implement `ViaTypeOf` and `ViaAttributeName` on attribute of dataclasses.
Fortunately, we already have special handling to generate models of dataclass constructors.
This diff improves it to properly translate `ViaTypeOf` and `ViaAttributeName` on attributes into the proper feature on the constructor argument.

Reviewed By: alexkassil

Differential Revision: D48398821

fbshipit-source-id: 878596025114584d0bc0365575dbf72736caba04