Ensure an invariant before and after each subtest in run_cache_test.py
Summary:
We want each subtest to begin and end with the same cache file that is
built from a "clean run" (i.e., from `Test.build_fresh_cache`), so that each
subtest begins with the proper setup (e.g., to avoid the wrong setup in
D48048219).

Such requirement is now explicated by a python decorator.

Reviewed By: alexkassil

Differential Revision: D48325444

fbshipit-source-id: a6b9af2547276b90b8c78982ccda65131a50e1f4