Enable saving and loading the maximum number of allowed overrides
Summary:
Since we should not reuse the override graph from the previous run when there
is a change in the max number of allowed overrides, we need to save the
previous run's max number of allowed overrides, to see if this configuration
has changed between two runs.

Alternatively, we could enable saving the whole taint configuration (which
contains the max number of allowed overrides), but there is no such need for
now.

It is expected that the new test in run_cache_test.py fails, due to not
invalidating the override graph (in the shared memory) from the previous run.
This will be fixed in the next diff. As a result, this diff needs to be landed
together with the next diff, to avoid breaking the cache test.

This diff also does a minor refactoring to run_cache_test.py, to avoid passing
the arguments around, by letting them be globals, so that the readability is
improved.

Reviewed By: arthaud

Differential Revision: D48195656

fbshipit-source-id: e3a96e1afb5ca4e6f29e78d77ae34108948bd2de