Properly return model errors during model query execution
Summary:
There are a few cases where we can find errors when generating models from model queries, even when the syntax is valid.
We currently just log those errors and continue the execution. For long running analyses, we will likely miss those.
To prevent this problem, let's properly propagate those errors and exit if the `--verify-dsl` flag was passed.

Reviewed By: tianhan0

Differential Revision: D48471717

fbshipit-source-id: 4549c9df3ed5b441eb9bbc4dae13c5faf05278b6