Raise model verification errors on model queries in integration tests
Summary:
We disabled those because some model queries do not return any result in integration tests.
Let's add different options for the different checks we perform, so we can only disable the empty result error in integration tests.

In the future, we might also want to be able to specify which error we want by providing flags to `pyre analyze`. Our current flags `--verify-models` and `--verify-dsl` are too generic.

Reviewed By: tianhan0

Differential Revision: D48471718

fbshipit-source-id: eaa510bf56f635c05d50229dde4b217741fe60ef