Fix crash in integration tests when trying to diff large inputs
Summary:
We use `diff -u` to show the difference between the expected tests and the actual tests.
Recently, we have added integration tests that produce some large output, which then makes `diff -u` crash.
There is a limit to how many bytes a command can receive via its arguments and environments, and we have reached that limit.

Let's catch that error and print a generic "diff -u failed" message for now.
In the future, we should probably use (or implement) a proper diffing library.

Reviewed By: tianhan0

Differential Revision: D48494620

fbshipit-source-id: 0a4390d165349eb1b2fcd9abb02c7323c13d77ab