Better documentation for taint broadening (or collapsing)
Summary:
# Problems
Our documentation for taint broadening is spread across the documentation.
This leads to some duplication, for instance in "Feature Annotations" and "Advanced Topics", where we describe collapsing twice.

# Solution
This diff moves most of the documentation about taint broadening under "Advanced Topics".
It adds sections about each type of broadening and describe which features are added.

Reviewed By: tianhan0

Differential Revision: D48523020

fbshipit-source-id: 7360165b317a3a317d053754b2b9ea08f5e96697