Add support for `_.parameter_name()`
Summary:
This diff implements `_.parameter_name()` which allows to define the `ParameterPath[]` or `ReturnPath[]` of `TaintInTaintOut[]` as the name of the parameter.

This will be useful when Pysa does not have access to the definition of a constructor, and that constructor just stores its parameters as attributes, similarly to a dataclass. For instance:
```python
class Point:
  def __init__(self, x, y, z):
    ... # missing code
```

Using `_.parameter_name()`, this can be properly modelled:
```
ModelQuery(
  name="point_constructor",
  find="methods",
  where=fully_qualified_name.equals("my_module.Point.__init__"),
  model=Parameters(TaintInTaintOut[LocalReturn, NoCollapse, ReturnPath[_.parameter_name()]],
)
```

This will automatically add the propagations (i.e TITO) from `x`, `y`, `z` to `self.x`, `self.y` and `self.z`.

Reviewed By: alexkassil

Differential Revision: D48648190

fbshipit-source-id: ed23505a102919d486f059a279d799b6be04ad5b