Provide the logical location of the taint annotation during parsing
Summary:
When parsing a taint annotation (for instance `TaintSource[UserControlled]`), we would like to know where the annotation is located (in a `def ..`, `ModelQuery`, etc.) to properly check for invalid models.

We currently pass two booleans `is_object_target` and `is_model_query`. This diff turns this into a proper type that distinguishes between all the different "logical" locations of a taint annotation.

In a follow-up, we will actually use this to check for invalid models.

Reviewed By: tianhan0

Differential Revision: D48684623

fbshipit-source-id: f12d44c1ef368251cdfbba1412e78907eb9e028f