Add utility function for attribute checks
Summary:
This is a small refactoring diff that adds a utility function that checks whether the annotation is for an attribute, since it is a common pattern.

We also swap "if/then/else" conditions by making the error branch the first "if" branch, which will make it easier in follow-ups to add error checks.

For instance:
```
if cond then
  Error "error"
// added in follow-up diffs
+ if new_cond then
+   Error "other error"
else
  Ok ...
```

Reviewed By: tianhan0

Differential Revision: D48686001

fbshipit-source-id: 5b38ca50b91aa3ae0ac2628338f235ac62dd29cb