Create a lightweight handle for SaveLoadSharedMemory.MakeKeyValue
Summary:
So that we can pass this lightweight handle (i.e., without the keys) around, to
improve performance. Otherwise, we will need to pass a copy of the handle to
each process, each of which analyzes a callable during the global fixpoint
computation. Such copy is costly when there are millions of keys.

Reviewed By: alexkassil

Differential Revision: D48661335

fbshipit-source-id: a1bc7c8f969be996a40a8865e3c653e084b4d66f