Cache the call graph
Summary:
This diff enables reusing the call graph from the previous run under certain
situations. That is, when all of the following conditions are met:
- there is no source code change (so that there is no change in the
  definitions)
- there is no change in the attribute targets (i.e., targets with attribute
  models)
- there is no override graph change
- there is no change in taint models with SkipAnalysis

Implementation-wise, this diff:
1. either reuses the call graph from the previous run, or computes a new one,
based on if the above conditions are all met.
2. addes `whole_program_call_graph` into Cache.AnalysisSetup, so that when
reusing the previous call graph, we can also reuse the previous
`whole_program_call_graph`.
3. adds tests for changes in taint models with SkipAnalysis and changes in the
definitions and attribute models.

Reviewed By: arthaud

Differential Revision: D48475257

fbshipit-source-id: 95bd68ff01dbe663adc7cbc4f7cee75e286fa5bc