Cache the global constants
Summary:
This diff enables reusing the global constants from the previous run, as long
as there is no source code change, which is manifested as loading the previous
shared memory and not resetting it.

In case the previous global constants are stale, we also don't need to
partially reset the shared memory, because the entire shared memory would have
been reset, when pysa detects source code changes.

Reviewed By: arthaud

Differential Revision: D48482099

fbshipit-source-id: e2285cd26514b6c1b0c5784cca0f70c2179359b1