Check for invalid model queries during parsing
Summary:
There are many cases where we successfully parse a taint annotation, but fail to generate a model when executing model queries.
This is bad user experience since model query generation happens later in the analysis and those errors are "soft errors" - meaning that the analysis continues unless `--verify-dsl` is provided. Those errors are likely to be ignored in production.
In this diff, we catch most errors during parsing to prevent those problems.

Reviewed By: dkgi

Differential Revision: D48736733

fbshipit-source-id: bdd00673aaccca25e74e1f9b8132eaf237cf2416