Add default formatting options
Summary:
This diff moves all formatting options (show_sources, show_sinks, show_features, etc.) in a dataclass.
It also introduces a set of default formatting options which only shows the essential information.
The reasoning is that in most cases, users only want to see source, sinks and tito with the relevant ports.

Reviewed By: tianhan0

Differential Revision: D48781847

fbshipit-source-id: ebaa616e251353ee99760cb5527b22f92f9e8c30