Fix ReadOnly[Self] bug
Summary:
See D48462009 for additional context

For Self variables, a transformation occurs as detailed here in Pep 637 https://peps.python.org/pep-0673/#use-in-method-signatures transforming

```
from typing import Self

class A:
    def a(self, b: float) -> Self:
        self.b = b
        return self
```

into

```
from typing import TypeVar

TA = TypeVar("TA", bound="A")

class A:
    def a(self: TA, b: float) -> TA:
        self.b = b
        return self
```

For models on self, usually it would be something like
```
module.A.b = TaintSource[Source]
```
which matches the naked implicit self, but not the typevar version.

So when we have `self: ReadOnly[Self]`, we both have a typevar bound to a class and a ReadOnly wrapper. This diff removes both iff both are present.

A potential followup would be to handle TypeVars and ReadOnly properly with pysa, both related to and not related to `self` usage.

Reviewed By: arthaud

Differential Revision: D48817989

fbshipit-source-id: 3f060caa077eeb70129ca923e0255ed15861c7e5