Avoid collecting global string literals from a different module
Summary:
# Context
For any given source file, we collect a map from the global variables within
this file to the string literals that are assigned to them (if any).

# This diff
However, sometimes the assigned global variables are imported from a different
module. In this case, we should not collect such entry, because
1. technically, in this case the string literal should "belong to" the module
that is imported,
2. it is probably fine to not support this usage, because this is a relatively
rare case,
3. if we collect this entry. then it may cause key collisions when merging maps
that are collected from different source files.

This diff also
- adds unit tests for collecting global constants, and
- adds function `pp` to module `Ast.Reference` (for easier debugging).

Reviewed By: alexkassil

Differential Revision: D48842467

fbshipit-source-id: d5e4e4b616b68be5d339f8c23f7b9c17c98a72f5