add indirection to pyre_server_options
Summary:
`pyre_server_options` is coupled to `start.create_server_arguments` which touches the file system. This makes testing `pyre_server_options` very difficult.

This diff's motivation comes from me attempting to add a field to `pyre_server_options` then realizing it was not necessary. I already factored this out for the test in the next diff and think it's worth keeping in some form. The next diff will include the unit test.

The only semantic change in this diff is that the binary location check happens after the server_arguments are created.

Reviewed By: grievejia

Differential Revision: D48791476

fbshipit-source-id: 23e76bb1947356dc20f4f32a7aa8933c18e19fd3