Extract functions that write the newline-delimited json files into a module
Summary:
When saving our results to a directory, we write taint models and issues into files using a format which we call "newline-delimited json".
We might want to use this format for other files (for instance, the call graph), but it is currently hardcoded in the taint reporting module.
Let's extract those functions into a proper module `NewlineDelimitedJson`.
This also takes care of sharding efficiently.

Reviewed By: alexkassil

Differential Revision: D48836083

fbshipit-source-id: adc4e1b24a8bbb3cf44c8252ee1557b8b3f7c668