Write the entire call graph to json
Summary:
When we write our call graph as json into a file, we only write the target map, which is basically of the form `"caller": ["callee1", "callee2"]`.
For debugging purposes, it could be useful to write down the whole call graph, including the precise location of each call and any unresolved calls (which are treated as obscure).
This diff implements that.

Reviewed By: tianhan0

Differential Revision: D48828455

fbshipit-source-id: e9c6f3032491b58d34e11235f502a0ab96b939d0