Add the full path of locations when a file is not within the repository
Summary:
Pysa only writes relative paths in the taint output and call graph files.
This is because the post processor (SAPP) and user interface can only show traces of files within the repository.
When a given file is outside the repository, we would just write `*`. This happens a lot for files in dependency libraries.

However, for debugging purposes, it can be useful to have the full path. This diff implements this.
To keep our output format backward compatible, we still provide the relative path under the `filename` key.
When the filename is `*`, we add an additional key `path` which contains the full path.

Reviewed By: alexkassil

Differential Revision: D48831370

fbshipit-source-id: 36bfa41607ebfd7c05e4370e08daaceaeb5fe41d