Allow to use `SkipObscure` in model queries
Summary:
We were currently not allowing `SkipObscure` in model queries. Let's fix that.

# Implementation
Until now, `SkipObscure` was basically a fake mode that was just translated into removing `Obscure`. This is the reason we had different handling of the modes for regular models and model queries. This diff fixes that by making it an explicit mode and having a special `join_user_modes` which handles the logic of merging conflicting modes.

Reviewed By: alexkassil, tianhan0

Differential Revision: D49016813

fbshipit-source-id: e6b267c8710e7d36c368ff58499d729a40cab282