Store the receiver class name instead of the whole type in the call graph
Summary:
We currently store the whole receiver type for each method call in the call graph.
In practice, we only really need the class name, for the implementation of class intervals, and for the special handling of named tuples.
This diff changes the call graph so we only store the receiver class name.
This should save us some memory usage. It will also produce smaller call graph output files.

Reviewed By: tianhan0

Differential Revision: D49102810

fbshipit-source-id: 2f9063d129207b762531ee9aae392bef6b21f93d