The receiver interval for static and class methods should be top
Summary:
We currently wrongly set the receiver class interval as the interval for the current class, but in practice the first parameter is an instance of `type`.
Let's just set it to top in those cases. Since most classes are of type `type` we won't be able to refine flows in cases.

Reviewed By: tianhan0

Differential Revision: D49157586

fbshipit-source-id: 382d9d432350c9c904b813baabea7c4a6d1e4018