Update Glass call routing
Summary:
Context:
Instagram-server database is going away for Glean, so we need to route all of our Glass calls to Glean's fbsource db.

Changes:
Glean requires us to specify the repository and have its symbol namespace to be relative to the repo. This typically means:
1. We have to redirect to the correct repo
2. Any calls with a filepath needs to have its param relative to the repo
3. Any calls where we feed in symbol needs its qualified name relative to the repo.

Fortunatley, all of our glean calls currently use documentSymbolListX, which fetches symbols relative to the repo for us; only things we have to address is 1 and 2.

I'm adding two helper methods to fetch fbsource and update the filepath relative to fbsource if the project root is instagram-server. This can be updated to support other repos (i.e. dataswarm-pipelines once they move into Glass).

Reviewed By: grievejia

Differential Revision: D49061370

fbshipit-source-id: c755e91f615ca992d7d69f595a7e44d38eda6d2a