add command to change global config strictness while maintaining local config strictness
Summary:
This diff creates the global-strictness command. From the docstring:

```
The global strictness command sets the global configuration's strictness field and adjusts the local configurations to have no difference.

For example, if a global configuration is strict and local configurations are non-strict, running the command with strict=False will set the global configuration to
strict=False and any unset local configurations will be set to strict=True. Alternatively, running with strict=True will set the global configuration to strict and
any unset configurations will be set to False.

After running this once, it is impossible to determine whether projects with a configuration set are intended to be strict / non-strict. A workaround would be to revert the
initial run.
```

Context:
This command is intended to be run on fbcode to default all new projects to strict=True and show current projects that strict=False.

Reviewed By: grievejia

Differential Revision: D49244320

fbshipit-source-id: 7544f014d2b8278be15a4ecb34cd7651401abc64